fix: logger is now static, reducing memory usage and test duration
