+++ src/handlers/logger-factory.ts
@@ -1,9 +1,9 @@
 import { LoggerService } from "./logger";
 
-export type ILoggerFactory = (name: string, logToFile?: boolean, logLevel?: string) => LoggerService;
+export type ILoggerFactory = (name: string) => LoggerService;
 
 export function LoggerFactory(): ILoggerFactory {
-  return (name: string, logToFile?: boolean, logLevel?: string) => {
-    return new LoggerService(name, logToFile, logLevel);
+  return (name: string) => {
+    return new LoggerService(name);
   };
 }
